backend/requirements.txt
Flask==2.3.3
Pillow==10.0.0
tensorflow>=2.16.1
flask-cors==5.0.0
numpy>=1.26.0
Werkzeug==3.0.1
gunicorn==21.2.0